requirements.txt
Django==4.0.6
djangorestframework==3.13.1
numpy==1.23.1
pandas==1.4.3
pytz==2022.1
